organized stuff into folders, added a new demo
added one function to calculator_utils.py
added a new demo folder
