--- big_df.csv
@@ -1,25 +1,0 @@
-,lat,lon,timezone,timezone_offset,dt,sunrise,sunset,temp,feels_like,pressure,humidity,dew_point,uvi,clouds,visibility,wind_speed,wind_deg,weather_id,weather_main,weather_description,weather_icon
-0,47.4,-121.5,America/Los_Angeles,-25200,1584403200,1584368146,1584411191,286.43,284.86,1018,40,273.13,,0,10000,0.0,0,800,Clear,clear sky,01d
-1,47.4,-121.5,America/Los_Angeles,-25200,1584406800,1584368146,1584411191,282.81,282.81,1018,42,270.83,,0,10000,0.0,0,800,Clear,clear sky,01d
-2,47.4,-121.5,America/Los_Angeles,-25200,1584410400,1584368146,1584411191,278.91,278.91,1018,83,276.25,,0,10000,0.0,0,800,Clear,clear sky,01d
-3,47.4,-121.5,America/Los_Angeles,-25200,1584414000,1584368146,1584411191,277.06,277.06,1017,87,275.1,,0,10000,0.0,0,800,Clear,clear sky,01n
-4,47.4,-121.5,America/Los_Angeles,-25200,1584417600,1584368146,1584411191,274.0,274.0,1017,80,271.2,,0,10000,0.0,0,800,Clear,clear sky,01n
-5,47.4,-121.5,America/Los_Angeles,-25200,1584421200,1584368146,1584411191,272.91,272.91,1017,83,270.67,,0,10000,0.0,0,800,Clear,clear sky,01n
-6,47.4,-121.5,America/Los_Angeles,-25200,1584424800,1584368146,1584411191,272.34,272.34,1016,84,270.25,,0,10000,0.0,0,800,Clear,clear sky,01n
-7,47.4,-121.5,America/Los_Angeles,-25200,1584428400,1584454425,1584497676,271.25,271.25,1015,86,269.46,,0,10000,0.0,0,800,Clear,clear sky,01n
-8,47.4,-121.5,America/Los_Angeles,-25200,1584432000,1584454425,1584497676,270.7,268.6,1014,86,268.91,,0,10000,1.5,80,800,Clear,clear sky,01n
-9,47.4,-121.5,America/Los_Angeles,-25200,1584435600,1584454425,1584497676,270.17,268.0,1013,87,268.53,,0,10000,1.5,0,800,Clear,clear sky,01n
-10,47.4,-121.5,America/Los_Angeles,-25200,1584439200,1584454425,1584497676,270.15,270.15,1012,93,269.29,,0,10000,0.0,0,800,Clear,clear sky,01n
-11,47.4,-121.5,America/Los_Angeles,-25200,1584442800,1584454425,1584497676,269.6,269.6,1012,94,268.87,,0,10000,0.0,0,800,Clear,clear sky,01n
-12,47.4,-121.5,America/Los_Angeles,-25200,1584446400,1584454425,1584497676,269.62,269.62,1011,89,268.25,,0,10000,0.0,0,800,Clear,clear sky,01n
-13,47.4,-121.5,America/Los_Angeles,-25200,1584450000,1584454425,1584497676,269.58,269.58,1011,91,268.47,,0,10000,0.0,0,800,Clear,clear sky,01n
-14,47.4,-121.5,America/Los_Angeles,-25200,1584453600,1584454425,1584497676,269.04,269.04,1010,91,267.93,,0,10000,0.0,0,800,Clear,clear sky,01n
-15,47.4,-121.5,America/Los_Angeles,-25200,1584457200,1584454425,1584497676,269.08,266.75,1010,91,267.97,,0,10000,1.5,80,800,Clear,clear sky,01d
-16,47.4,-121.5,America/Los_Angeles,-25200,1584460800,1584454425,1584497676,269.66,267.42,1010,90,268.42,,0,10000,1.5,0,800,Clear,clear sky,01d
-17,47.4,-121.5,America/Los_Angeles,-25200,1584464400,1584454425,1584497676,282.07,282.07,1010,44,270.79,,0,10000,0.0,0,800,Clear,clear sky,01d
-18,47.4,-121.5,America/Los_Angeles,-25200,1584468000,1584454425,1584497676,277.36,275.03,1009,72,272.81,,0,10000,2.6,0,800,Clear,clear sky,01d
-19,47.4,-121.5,America/Los_Angeles,-25200,1584471600,1584454425,1584497676,283.1,283.1,1009,58,275.23,,0,10000,0.0,0,800,Clear,clear sky,01d
-20,47.4,-121.5,America/Los_Angeles,-25200,1584475200,1584454425,1584497676,283.7,282.22,1009,54,274.79,,0,10000,2.1,0,800,Clear,clear sky,01d
-21,47.4,-121.5,America/Los_Angeles,-25200,1584478800,1584454425,1584497676,285.25,283.85,1008,51,275.43,,0,10000,1.5,0,800,Clear,clear sky,01d
-22,47.4,-121.5,America/Los_Angeles,-25200,1584482400,1584454425,1584497676,285.32,283.92,1007,51,275.49,,0,10000,1.5,0,800,Clear,clear sky,01d
-23,47.4,-121.5,America/Los_Angeles,-25200,1584486000,1584454425,1584497676,286.39,284.89,1007,43,274.09,,0,10000,0.36,351,800,Clear,clear sky,01d
