--- history_bulk_example_csv.csv
@@ -1,49 +1,0 @@
-dt,dt_iso,timezone,city_name,lat,lon,temp,feels_like,temp_min,temp_max,pressure,sea_level,grnd_level,humidity,wind_speed,wind_deg,rain_1h,rain_3h,snow_1h,snow_3h,clouds_all,weather_id,weather_main,weather_description,weather_icon
-1593561600,2020-07-01 00:00:00 +0000 UTC,3600,London,51.507351,-0.127758,289.89,287.15,288.468,291.15,1004,,,77,5.1,230,,,,,90,804,Clouds,overcast clouds,04n
-1593565200,2020-07-01 01:00:00 +0000 UTC,3600,London,51.507351,-0.127758,289.27,287.7,287.588,290.37,1004,,,82,3.6,230,,,,,90,500,Rain,light rain,10n
-1593568800,2020-07-01 02:00:00 +0000 UTC,3600,London,51.507351,-0.127758,288.97,285.66,287.257,290.15,1003,,,72,5.1,240,,,,,90,804,Clouds,overcast clouds,04n
-1593572400,2020-07-01 03:00:00 +0000 UTC,3600,London,51.507351,-0.127758,288.33,285.48,287.037,289.26,1003,,,77,4.6,240,,,,,75,803,Clouds,broken clouds,04n
-1593576000,2020-07-01 04:00:00 +0000 UTC,3600,London,51.507351,-0.127758,287.75,285.09,285.93,289.15,1003,,,77,4.1,240,,,,,75,803,Clouds,broken clouds,04d
-1593579600,2020-07-01 05:00:00 +0000 UTC,3600,London,51.507351,-0.127758,287.19,284.3,285.37,289.1,1003,,,82,4.6,230,,,,,0,800,Clear,sky is clear,01d
-1593583200,2020-07-01 06:00:00 +0000 UTC,3600,London,51.507351,-0.127758,288.27,285.12,287.04,289.15,1003,,,72,4.6,250,,,,,0,800,Clear,sky is clear,01d
-1593586800,2020-07-01 07:00:00 +0000 UTC,3600,London,51.507351,-0.127758,289.4,285.91,287.3,290.93,1003,,,67,5.1,250,,,,,0,800,Clear,sky is clear,01d
-1593590400,2020-07-01 08:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.12,287.05,289.15,292.59,1004,,,63,6.2,250,,,,,20,801,Clouds,few clouds,02d
-1593594000,2020-07-01 09:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.94,288.44,290.15,293.15,1004,,,56,5,240,,,,,40,802,Clouds,scattered clouds,03d
-1593597600,2020-07-01 10:00:00 +0000 UTC,3600,London,51.507351,-0.127758,292.7,289.39,290.93,294.26,1004,,,56,5,240,,,,,75,803,Clouds,broken clouds,04d
-1593601200,2020-07-01 11:00:00 +0000 UTC,3600,London,51.507351,-0.127758,293.15,289.83,291.15,294.82,1004,,,52,4.74,238,,,,,40,802,Clouds,scattered clouds,03d
-1593604800,2020-07-01 12:00:00 +0000 UTC,3600,London,51.507351,-0.127758,293.13,289.91,291.15,294.82,1004,,,52,4.6,250,,,,,75,803,Clouds,broken clouds,04d
-1593608400,2020-07-01 13:00:00 +0000 UTC,3600,London,51.507351,-0.127758,293.51,290.03,292.04,294.82,1004,,,52,5.1,240,,1,,,0,500,Rain,light rain,10d
-1593612000,2020-07-01 14:00:00 +0000 UTC,3600,London,51.507351,-0.127758,292.79,288.91,290.902,295.15,1005,,,49,5.1,260,,0.31,,,100,500,Rain,light rain,10d
-1593615600,2020-07-01 15:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.68,288.75,289.82,293.15,1006,,,72,5.7,250,,0.31,,,100,500,Rain,light rain,10d
-1593619200,2020-07-01 16:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.68,288.75,289.26,293.15,1005,,,72,5.7,230,,0.31,,,100,500,Rain,light rain,10d
-1593622800,2020-07-01 17:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.89,289.79,289.26,293.71,1006,,,72,4.6,230,,,,,0,800,Clear,sky is clear,01d
-1593626400,2020-07-01 18:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.26,287.92,289.26,292.59,1005,,,68,5.7,250,,,,,0,800,Clear,sky is clear,01d
-1593630000,2020-07-01 19:00:00 +0000 UTC,3600,London,51.507351,-0.127758,290.56,287.02,288.15,293.15,1006,,,68,5.7,250,,,,,75,803,Clouds,broken clouds,04d
-1593633600,2020-07-01 20:00:00 +0000 UTC,3600,London,51.507351,-0.127758,289.48,286.2,287.15,291.48,1006,,,77,5.7,250,,,,,79,500,Rain,light rain,10d
-1593637200,2020-07-01 21:00:00 +0000 UTC,3600,London,51.507351,-0.127758,289.05,287.46,287.15,290.37,1007,,,77,3.1,240,,,,,79,803,Clouds,broken clouds,04n
-1593640800,2020-07-01 22:00:00 +0000 UTC,3600,London,51.507351,-0.127758,289.03,287.38,287.15,290.93,1007,,,82,3.6,230,,,,,79,803,Clouds,broken clouds,04n
-1593644400,2020-07-01 23:00:00 +0000 UTC,3600,London,51.507351,-0.127758,288.51,286.42,287.04,290.15,1007,,,77,3.6,240,,,,,0,800,Clear,sky is clear,01n
-1593648000,2020-07-02 00:00:00 +0000 UTC,3600,London,51.507351,-0.127758,288.06,285.42,285.93,289.82,1007,,,82,4.6,250,,,,,40,802,Clouds,scattered clouds,03n
-1593651600,2020-07-02 01:00:00 +0000 UTC,3600,London,51.507351,-0.127758,287.04,282.34,284.82,289.26,1007,,,63,5.7,270,,,,,20,801,Clouds,few clouds,02n
-1593655200,2020-07-02 02:00:00 +0000 UTC,3600,London,51.507351,-0.127758,286,282.4,283.376,288.15,1007,,,67,4.1,240,,,,,77,803,Clouds,broken clouds,04n
-1593658800,2020-07-02 03:00:00 +0000 UTC,3600,London,51.507351,-0.127758,285.43,282.84,283.15,287.15,1008,,,76,3.1,240,,,,,77,803,Clouds,broken clouds,04n
-1593662400,2020-07-02 04:00:00 +0000 UTC,3600,London,51.507351,-0.127758,285.12,281.4,283.15,287.15,1007,,,76,4.6,240,,,,,77,803,Clouds,broken clouds,04d
-1593666000,2020-07-02 05:00:00 +0000 UTC,3600,London,51.507351,-0.127758,285.29,281.89,283.15,287.639,1008,,,82,4.6,240,,1,,,0,500,Rain,light rain,10d
-1593669600,2020-07-02 06:00:00 +0000 UTC,3600,London,51.507351,-0.127758,286.42,283.32,284.82,287.639,1008,,,82,4.6,250,,1,,,0,500,Rain,light rain,10d
-1593673200,2020-07-02 07:00:00 +0000 UTC,3600,London,51.507351,-0.127758,287.98,284.41,286.15,289.82,1009,,,72,5.1,240,,1,,,0,500,Rain,light rain,10d
-1593676800,2020-07-02 08:00:00 +0000 UTC,3600,London,51.507351,-0.127758,288.96,284.93,287.04,291.134,1009,,,67,5.7,260,,,,,75,803,Clouds,broken clouds,04d
-1593680400,2020-07-02 09:00:00 +0000 UTC,3600,London,51.507351,-0.127758,289.48,285.59,288.15,291.134,1009,,,67,5.7,250,,,,,75,803,Clouds,broken clouds,04d
-1593684000,2020-07-02 10:00:00 +0000 UTC,3600,London,51.507351,-0.127758,290.51,286.96,288.71,291.48,1009,,,68,5.7,250,,,,,75,803,Clouds,broken clouds,04d
-1593687600,2020-07-02 11:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.32,288.08,289.26,293.71,1010,,,63,5.1,240,0.51,,,,0,500,Rain,light rain,10d
-1593691200,2020-07-02 12:00:00 +0000 UTC,3600,London,51.507351,-0.127758,290.88,287.43,288.15,293.238,1010,,,68,5.7,240,,1,,,0,500,Rain,light rain,10d
-1593694800,2020-07-02 13:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.16,287.79,289.15,293.238,1011,,,68,5.7,250,,,,,0,500,Rain,light rain,10d
-1593698400,2020-07-02 14:00:00 +0000 UTC,3600,London,51.507351,-0.127758,292.74,289.5,290.15,295.93,1011,,,68,6.2,250,0.25,,,,12,500,Rain,light rain,10d
-1593702000,2020-07-02 15:00:00 +0000 UTC,3600,London,51.507351,-0.127758,293.44,289.45,290.93,296.48,1011,,,60,6.7,250,,1.44,,,12,500,Rain,light rain,10d
-1593705600,2020-07-02 16:00:00 +0000 UTC,3600,London,51.507351,-0.127758,293.6,290.36,291.48,295.273,1011,,,60,5.7,260,,1.44,,,12,500,Rain,light rain,10d
-1593709200,2020-07-02 17:00:00 +0000 UTC,3600,London,51.507351,-0.127758,292.6,288.57,290.1,294.15,1012,,,72,7.7,300,,,,,0,500,Rain,light rain,10d
-1593712800,2020-07-02 18:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.67,293.08,289.26,294.15,1013,,,82,0.5,0,,,,,0,500,Rain,light rain,10d
-1593716400,2020-07-02 19:00:00 +0000 UTC,3600,London,51.507351,-0.127758,291.09,290.82,288.71,293.7,1013,,,82,2.6,250,,,,,0,500,Rain,light rain,10d
-1593720000,2020-07-02 20:00:00 +0000 UTC,3600,London,51.507351,-0.127758,290.25,290.85,287.15,292.04,1013,,,88,1.5,220,,0.5,,,5,500,Rain,light rain,10d
-1593723600,2020-07-02 21:00:00 +0000 UTC,3600,London,51.507351,-0.127758,290.03,290.13,287.15,292.04,1014,,,88,2.1,240,,0.5,,,5,500,Rain,light rain,10n
-1593727200,2020-07-02 22:00:00 +0000 UTC,3600,London,51.507351,-0.127758,289.96,290.46,287.15,292.04,1015,,,88,1.5,210,,,,,20,801,Clouds,few clouds,02n
-1593730800,2020-07-02 23:00:00 +0000 UTC,3600,London,51.507351,-0.127758,289.11,288.09,287.003,290.93,1015,,,88,3.25,277,,,,,75,803,Clouds,broken clouds,04n
